// KROW/backend/src/models/Curriculum.ts
export interface Curriculum {
  id_curriculum: number;
  usuario_id: number;
  perfil_profesional: string | null;
  campo_laboral: string | null;
  campo_estudiantil: string | null;
  fortalezas: string | null;
  debilidades: string | null;
  idiomas: string | null;
  habilidades: string | null;
  certificaciones: string | null;
  portafolio: string | null;
  fecha_actualizacion: Date;
}